requirements.txt
aioboto3
boto3
cachetools
python-dotenv